V0.9.2
The user can now change between dark and light mode, with the default being dark mode. The key bind is c.
The user can also toggle whether they display the FPS or not. the key bind is f and F3.
The next thing I intend on doing is adding a way for the user to change the screen size while they are playing the game.